# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EnergieGo API",
        "description": "Documentation OpenAPI pour l'API EnergieGo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/clients": {
            "get": {
                "summary": "Lister les clients",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un client",
                "operationId": "5e364247262acc76137f02b2b177d636",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "summary": "Afficher un client",
                "operationId": "4c3c7071ea1133f04bacdbc70f8aa230",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un client",
                "operationId": "4b22619962e2decc103e8961512f3cb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un client",
                "operationId": "c929f14f421a80a896c6fe116ffdc317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/commandes": {
            "get": {
                "summary": "Lister les commandes",
                "operationId": "6a11b062792d6b0aed3816592a638d6c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Commande"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une commande",
                "operationId": "5528848ef74479f1b283ec1da16083e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commande"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commande"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/commandes/{id}": {
            "get": {
                "summary": "Afficher une commande",
                "operationId": "700d31135d2d0bcc8efcda92ca45ae8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commande"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une commande",
                "operationId": "06c0eb23c4cdba5a384e2a291834dda3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commande"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commande"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une commande",
                "operationId": "c9022bad4403150c0e68ebfbfb507a69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/lignes-commande": {
            "get": {
                "summary": "Lister les lignes de commande",
                "operationId": "0bd92e4d13ea81ef3e9327227d48d867",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LigneCommande"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une ligne de commande",
                "operationId": "5e3f2ba036e53943ba062adfe6e8124d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LigneCommande"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LigneCommande"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lignes-commande/{id}": {
            "get": {
                "summary": "Afficher une ligne de commande",
                "operationId": "41590841a0c6f13115279f697514bc25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LigneCommande"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une ligne de commande",
                "operationId": "eef7ffe11a7a7dceb9403e4761348b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LigneCommande"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LigneCommande"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une ligne de commande",
                "operationId": "03fc14b4836f4a9e701570d8ad79dfa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "required": [
                    "id_utilisateur",
                    "adresse",
                    "telephone"
                ],
                "properties": {
                    "id_utilisateur": {
                        "type": "integer",
                        "example": 1
                    },
                    "adresse": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Commande": {
                "required": [
                    "id_paiement",
                    "id_utilisateur",
                    "dateCommande",
                    "statut",
                    "total"
                ],
                "properties": {
                    "id_commande": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_paiement": {
                        "type": "integer"
                    },
                    "id_utilisateur": {
                        "type": "integer"
                    },
                    "dateCommande": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "statut": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "LigneCommande": {
                "required": [
                    "id_produit",
                    "id_commande",
                    "quantite",
                    "sousTotal"
                ],
                "properties": {
                    "id_ligne_commande": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_produit": {
                        "type": "integer"
                    },
                    "id_commande": {
                        "type": "integer"
                    },
                    "quantite": {
                        "type": "integer"
                    },
                    "sousTotal": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
        }
    }
}